Debug Update
- Added debug class with macro functions to print to console log and output file
- Added debug log to error cases through ecs project

+++ EntityComponentSystem/ECS/include/Parser.h
@@ -23,7 +23,7 @@ namespace ECS
 				return;
 			}
 
-			int entityCount = world->m_enityManager->m_entityCounter;
+			uint64_t entityCount = world->m_enityManager->m_entityCounter;
 
 			for ( int i = 1; i <= entityCount; i++ )
 			{
